Lau - DEC - Indicadores

+++ Docs/07-Implementacion/Source/trunk/EDUAR_actual/EDUAR/EDUAR_DataTransferObject/DEC/ConfigFuncionPreferencia.cs
@@ -10,6 +10,6 @@ namespace EDUAR_Entities
 		public int idIndicador { get; set; }
 		public int idFuncionPreferencia { get; set; }
 		public int idValorFuncionPreferencia { get; set; }
-		public float valorDefault { get; set; }
+		public decimal valorDefault { get; set; }
 	}
 }
